Rename interaction -> interactive
