+++ src/game_logic/entity_factory.cpp
@@ -50,16 +50,16 @@
 #include "game_logic/dynamic_geometry_components.hpp"
 #include "game_logic/effect_actor_components.hpp"
 #include "game_logic/effect_components.hpp"
-#include "game_logic/interaction/blowing_fan.hpp"
-#include "game_logic/interaction/elevator.hpp"
-#include "game_logic/interaction/enemy_radar.hpp"
-#include "game_logic/interaction/force_field.hpp"
-#include "game_logic/interaction/locked_door.hpp"
-#include "game_logic/interaction/missile.hpp"
-#include "game_logic/interaction/respawn_checkpoint.hpp"
-#include "game_logic/interaction/sliding_door.hpp"
-#include "game_logic/interaction/super_force_field.hpp"
-#include "game_logic/interaction/tile_burner.hpp"
+#include "game_logic/interactive/blowing_fan.hpp"
+#include "game_logic/interactive/elevator.hpp"
+#include "game_logic/interactive/enemy_radar.hpp"
+#include "game_logic/interactive/force_field.hpp"
+#include "game_logic/interactive/locked_door.hpp"
+#include "game_logic/interactive/missile.hpp"
+#include "game_logic/interactive/respawn_checkpoint.hpp"
+#include "game_logic/interactive/sliding_door.hpp"
+#include "game_logic/interactive/super_force_field.hpp"
+#include "game_logic/interactive/tile_burner.hpp"
 #include "game_logic/item_container.hpp"
 #include "game_logic/trigger_components.hpp"
