+++ src/game_logic/ingame_systems.hpp
@@ -39,9 +39,9 @@
 #include "game_logic/debugging_system.hpp"
 #include "game_logic/dynamic_geometry_system.hpp"
 #include "game_logic/effects_system.hpp"
-#include "game_logic/interaction/elevator.hpp"
-#include "game_logic/interaction/enemy_radar.hpp"
-#include "game_logic/interaction/sliding_door.hpp"
+#include "game_logic/interactive/elevator.hpp"
+#include "game_logic/interactive/enemy_radar.hpp"
+#include "game_logic/interactive/sliding_door.hpp"
 #include "game_logic/item_container.hpp"
 #include "game_logic/player.hpp"
 #include "game_logic/player/damage_system.hpp"
