+++ src/game_logic/interactive/locked_door.hpp
