+++ src/game_logic/interactive/respawn_checkpoint.hpp
