+++ src/game_logic/interactive/sliding_door.cpp
