+++ src/game_logic/interactive/sliding_door.hpp
