+++ src/game_logic/interactive/tile_burner.hpp
