+++ src/game_logic/player_interaction_system.cpp
@@ -25,8 +25,8 @@
 #include "game_logic/collectable_components.hpp"
 #include "game_logic/damage_components.hpp"
 #include "game_logic/entity_factory.hpp"
-#include "game_logic/interaction/force_field.hpp"
-#include "game_logic/interaction/locked_door.hpp"
+#include "game_logic/interactive/force_field.hpp"
+#include "game_logic/interactive/locked_door.hpp"
 #include "game_logic/player.hpp"
 #include "loader/duke_script_loader.hpp"
 #include "loader/resource_loader.hpp"
